pyproject.toml
# Copyright (c) Meta Platforms, Inc. and affiliates.
# All rights reserved.
#
# This source code is licensed under the BSD-style license found in the
# LICENSE file in the root directory of this source tree.

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-bulkTrade"
version = "0.1.0"
description = "Bulktrade environment for OpenEnv"
requires-python = ">=3.10"
dependencies = [
    "openenv-core[core]>=0.2.1",
    "numpy>=1.26.0",
    "pandas>=2.2.0",
    "yfinance>=0.2.54",
    "openai>=1.40.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
# Server entry point - enables running via: uv run --project . server
# or: python -m bulkTrade.server.app
server = "bulkTrade.server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["bulkTrade", "bulkTrade.server"]
package-dir = { "bulkTrade" = ".", "bulkTrade.server" = "server" }